Fix signs on velocity vector and turn coordinator.
Changes the sign of alpha and beta when passed on to the QFI PFD so that
the velocity vector is correctly displayed compared to the X-Plane HUD's
flight path vector.

Additionally, change the sign of the turn coordinator/slip skid update,
also to match X-Plane's HUD.

Some other changes are just trivial removal of trailing whitespace done
automagically by my vimrc.